# openapi.yaml
openapi: 3.0.3
info:
  title: PadelZGZ API
  description: API REST para gestión de clubes de pádel, reservas de pistas, torneos e inscripciones en Zaragoza.
  version: 1.0.0
  contact:
    name: PadelZGZ
    email: info@padelzgz.com
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Club:
      type: object
      required: [nombre, ciudad]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nombre:
          type: string
          example: "Club Pádel Zaragoza"
        ciudad:
          type: string
          example: "Zaragoza"
        direccion:
          type: string
          example: "Calle Mayor 10"
        telefono:
          type: string
          pattern: "^[0-9]{9}$"
          example: "976123456"
        email:
          type: string
          format: email
          example: "info@clubpadel.com"
        activo:
          type: boolean
          example: true
        fechaApertura:
          type: string
          format: date
          example: "2020-03-15"
    Pista:
      type: object
      required: [numero, tipo]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        numero:
          type: integer
          minimum: 1
          example: 3
        tipo:
          type: string
          enum: [PANORAMICA, CRISTAL, MURO]
          example: "CRISTAL"
        interior:
          type: boolean
          example: true
        precioHora:
          type: number
          format: float
          example: 12.50
        activa:
          type: boolean
          example: true
        superficie:
          type: string
          enum: [CESPED, MOQUETA, HORMIGON]
          example: "CESPED"
...
    Torneo:
      type: object
      required: [nombre, fechaInicio]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nombre:
          type: string
          example: "Torneo Verano ZGZ 2025"
        descripcion:
          type: string
          example: "Torneo de verano para jugadores de nivel intermedio y avanzado"
        fechaInicio:
          type: string
          format: date
          example: "2025-07-01"
        fechaFin:
          type: string
          format: date
          example: "2025-07-07"
        maxParticipantes:
          type: integer
          minimum: 2
          example: 16
        inscripcionAbierta:
          type: boolean
          example: true
        precioInscripcion:
          type: number
          format: float
          example: 25.00
...
    ClubPistasResumen:
      type: object
      properties:
        clubId:
          type: integer
          format: int64
        nombreClub:
          type: string
        ciudad:
          type: string
        totalPistas:
          type: integer
        pistasActivas:
          type: integer
        pistas:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
              numero:
                type: integer
              tipo:
                type: string
              interior:
                type: boolean
              precioHora:
                type: number
              valoracionMedia:
                type: number
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 404
        message:
          type: string
          example: "Club no encontrado: 99"
        errors:
          type: object
          additionalProperties:
            type: string
paths:
...
  /clubs:
    get:
      tags: [Clubs]
      summary: Listar clubs
      parameters:
        - name: ciudad
          in: query
          schema:
            type: string
          example: "Zaragoza"
        - name: activo
          in: query
          schema:
            type: boolean
          example: true
      responses:
        '200':
          description: Lista de clubs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Club'
              examples:
                clubs_zaragoza:
                  summary: Clubs en Zaragoza
                  value:
                    - id: 1
                      nombre: "Club Pádel Zaragoza"
                      ciudad: "Zaragoza"
                      activo: true
                clubs_vacio:
                  summary: Sin resultados
                  value: []
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Clubs]
      summary: Crear un club
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Club'
            examples:
              club_nuevo:
                summary: Club nuevo en Zaragoza
                value:
                  nombre: "Pádel Norte ZGZ"
                  ciudad: "Zaragoza"
                  telefono: "976000111"
                  activo: true
                  fechaApertura: "2024-09-01"
      responses:
        '201':
          description: Club creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                error_validacion:
                  summary: Falta nombre
                  value:
                    code: 400
                    message: "Error de validación"
                    errors:
                      nombre: "El nombre es obligatorio"
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /clubs/{id}:
    get:
      tags: [Clubs]
      summary: Obtener un club por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Club encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '404':
          description: Club no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_found:
                  value: { code: 404, message: "Club no encontrado: 99" }
        '500':
          description: Error interno

    put:
      tags: [Clubs]
      summary: Modificar un club completo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Club'
      responses:
        '200':
          description: Club actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '400':
          description: Datos inválidos
        '404':
          description: Club no encontrado
        '500':
          description: Error interno

    patch:
      tags: [Clubs]
      summary: Modificar campos de un club parcialmente
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Club'
            examples:
              patch_nombre:
                summary: Cambiar solo el nombre
                value: { nombre: "Club Pádel ZGZ Premium" }
              patch_activo:
                summary: Desactivar club
                value: { activo: false }
      responses:
        '200':
          description: Club parcialmente actualizado
        '404':
          description: Club no encontrado
        '500':
          description: Error interno

    delete:
      tags: [Clubs]
      summary: Eliminar un club
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Club eliminado
        '404':
          description: Club no encontrado
        '500':
          description: Error interno
  /clubs/{id}/pistas-resumen:
    get:
      tags: [Clubs]
      summary: Resumen de pistas de un club con valoración media (DTO relacional)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Resumen de pistas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClubPistasResumen'
        '404':
          description: Club no encontrado
        '500':
          description: Error interno
  /clubs/mas-activos:
    get:
      tags: [Clubs]
      summary: Clubs con más pistas activas (SQL nativa)
      responses:
        '200':
          description: Lista ordenada por número de pistas activas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Club'
        '500':
          description: Error interno
  /pistas:
    get:
      tags: [Pistas]
      summary: Listar pistas con filtros opcionales
      parameters:
        - name: tipo
          in: query
          schema:
            type: string
        - name: interior
          in: query
          schema:
            type: boolean
        - name: activa
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de pistas
        '500':
          description: Error interno
  /pistas/club/{clubId}:
    post:
      tags: [Pistas]
      summary: Crear pista en un club
      security:
        - bearerAuth: []
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pista'
      responses:
        '201':
          description: Pista creada
        '400':
          description: Datos inválidos
        '404':
          description: Club no encontrado
        '500':
          description: Error interno
  /pistas/por-precio:
    get:
      tags: [Pistas]
      summary: Pistas en rango de precio (SQL nativa)
      parameters:
        - name: min
          in: query
          required: true
          schema:
            type: number
        - name: max
          in: query
          required: true
          schema:
            type: number
      responses:
        '200':
          description: Pistas en el rango indicado
        '500':
          description: Error interno
  /pistas/mejor-valoradas:
    get:
      tags: [Pistas]
      summary: Pistas con puntuación media mínima (SQL nativa)
      parameters:
        - name: minPuntuacion
          in: query
          schema:
            type: number
            default: 4
      responses:
        '200':
          description: Pistas mejor valoradas
        '500':
          description: Error interno
...
  /torneos/club/{clubId}:
    post:
      tags: [Torneos]
      summary: Crear torneo en un club
      security:
        - bearerAuth: []
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Torneo'
      responses:
        '201':
          description: Torneo creado
        '400':
          description: Datos inválidos
        '404':
          description: Club no encontrado
        '500':
          description: Error interno
...